+++ sdk/eventgrid/eventgrid-namespaces/karma.conf.js
@@ -3,8 +3,8 @@
 
 // https://github.com/karma-runner/karma-chrome-launcher
 process.env.CHROME_BIN = require("puppeteer").executablePath();
-require("dotenv").config();
 const { relativeRecordingsPath } = require("@azure-tools/test-recorder");
+require("dotenv").config();
 
 process.env.RECORDINGS_RELATIVE_PATH = relativeRecordingsPath();
 
@@ -29,10 +29,7 @@ module.exports = function (config) {
     ],
 
     // list of files / patterns to load in the browser
-    files: [
-      "dist-test/index.browser.js",
-      { pattern: "dist-test/index.browser.js.map", type: "html", included: false, served: true },
-    ],
+    files: ["dist-test/index.browser.js"],
 
     // list of files / patterns to exclude
     exclude: [],
@@ -43,20 +40,22 @@ module.exports = function (config) {
       "**/*.js": ["sourcemap", "env"],
       // IMPORTANT: COMMENT following line if you want to debug in your browsers!!
       // Preprocess source file to calculate code coverage, however this will make source file unreadable
-      //"dist-test/index.browser.js": ["coverage"]
+      // "dist-test/index.browser.js": ["coverage"]
     },
 
-    // inject following environment values into browser testing with window.__env__
-    // environment values MUST be exported or set with same console running "karma start"
-    // https://www.npmjs.com/package/karma-env-preprocessor
     envPreprocessor: [
       "TEST_MODE",
-      "WIDGET_SERVICE_ENDPOINT",
+      "RECORDINGS_RELATIVE_PATH",
+      "AZURE_TENANT_ID",
       "AZURE_CLIENT_ID",
       "AZURE_CLIENT_SECRET",
-      "AZURE_TENANT_ID",
-      "RECORDINGS_RELATIVE_PATH",
+      "EVENT_GRID_NAMESPACES_ENDPOINT",
+      "EVENT_GRID_NAMESPACES_KEY",
+      "EVENT_SUBSCRIPTION_NAME",
+      "TOPIC_NAME",
+      "MAX_DELIVERY_COUNT",
     ],
+
     // test results reporter to use
     // possible values: 'dots', 'progress'
     // available reporters: https://npmjs.org/browse/keyword/karma-reporter
@@ -96,22 +95,20 @@ module.exports = function (config) {
     // enable / disable watching file and executing tests whenever any file changes
     autoWatch: false,
 
-    // start these browsers
-    // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
-    // 'ChromeHeadless', 'Chrome', 'Firefox', 'Edge', 'IE'
-    browsers: ['ChromeHeadlessNoSandbox'],
-
+    // --no-sandbox allows our tests to run in Linux without having to change the system.
+    // --disable-web-security allows us to authenticate from the browser without having to write tests using interactive auth, which would be far more complex.
+    browsers: ["ChromeHeadlessNoSandbox"],
     customLaunchers: {
       ChromeHeadlessNoSandbox: {
-        base: 'ChromeHeadless',
-        flags: ['--no-sandbox', "--disable-web-security"]
-      }
+        base: "ChromeHeadless",
+        flags: ["--no-sandbox", "--disable-web-security"],
+      },
     },
 
     // Continuous Integration mode
     // if true, Karma captures browsers, runs the tests and exits
-    singleRun: true,
-    
+    singleRun: false,
+
     // Concurrency level
     // how many browser should be started simultaneous
     concurrency: 1,